Rename localEvents
