+++ _common/gameModel/libs/ClientScript/game/common/impacts/implementation/activationImpact.ts
@@ -9,7 +9,7 @@ import { ImpactBase } from '../impact';
 import {
   ChangeActivableStatusLocalEvent,
   ChangeMapActivableStatusLocalEvent,
-} from '../../localEvents/localEventBaseActivable';
+} from '../../localEvents/localEventActivable';
 
 export type ActivationOperator = 'activate' | 'deactivate';
