+++ _common/gameModel/libs/ClientScript/game/common/impacts/implementation/choiceEffectSelectionImpact.ts
@@ -2,7 +2,7 @@ import { Uid } from '../../interfaces';
 import { LocalEventBase, SourceType } from '../../localEvents/localEventBase';
 import { MainSimulationState } from '../../simulationState/mainSimulationState';
 import { ImpactBase } from '../impact';
-import { SelectChoiceEffectLocalEvent } from '../../localEvents/localEventBaseActivable';
+import { SelectChoiceEffectLocalEvent } from '../../localEvents/localEventActivable';
 
 export interface ChoiceEffectSelectionImpact extends ImpactBase {
   type: 'effectSelection';
