+++ _common/gameModel/libs/ClientScript/game/common/localEvents/localEventResources.ts
